test cache change config

+++ src/algorithms/quantization/hqq.py
@@ -63,9 +63,8 @@ def quantize_hqq(model_name, num_bits=None, dynamic_config=False, save_model=Fal
     hqq_model = AutoModelForCausalLM.from_pretrained(
         model_name,
         quantization_config=hqq_config, 
-        # torch_dtype=torch.float16,
-        device_map=device,
-        # force_download=True
+        torch_dtype=torch.float16,
+        device_map=device
     )
     end_time = time.time()  # End time measurement
     hqq_model.QUANT_TIME = end_time - start_time
